made git graph use SVG to reduce latency to unnoticeable levels
