+++ src/app/components/git-graph-canvas/git-graph-canvas.component.scss
@@ -1,4 +1,4 @@
-canvas {
+canvas, svg {
   top: 0;
   left: 0;
 }
