+++ src/styles.scss
@@ -32,6 +32,7 @@ html, body {
   -moz-border-radius: $border-radius;
   border-radius: $border-radius;
 }
+
 .dropdown-menu {
   box-shadow: $material-dialog-shadow;
   border: none;
@@ -121,9 +122,11 @@ body {
   margin: auto;
   max-height: 60%;
   box-shadow: $material-dialog-shadow;
+
   .modal-content {
     display: flex;
     flex-direction: column;
+
     .modal-body {
       overflow-y: auto;
       flex-grow: 1;
@@ -247,4 +250,10 @@ body {
 
 .table td {
   vertical-align: middle;
+}
+
+svg text {
+  font-family: "Font Awesome 5 Free";
+  font-size: .9em;
+  font-weight: 900;
 }
